Moved login to plugin and register to server.
